add embeddings tests, cleanup
