+++ src/cltkv1/embeddings/__init__.py
@@ -1,4 +1,4 @@
 """Init for ``cltkv1.embeddings``."""
 
-# from .embeddings import *
-# from .word import *
+from .embeddings import *
+from .processes import *
