+++ src/cltkv1/utils/utils.py
@@ -5,8 +5,6 @@ import sys
 from contextlib import contextmanager
 from typing import Any, Dict, List, Optional
 
-# from cltkv1 import __cltk_data_dir__
-
 
 def file_exists(file_path: str, is_dir: bool = False) -> bool:
     """Try to expand `~/` and check if a file or dir exists.
@@ -99,23 +97,44 @@ def suppress_stdout():
             sys.stdout = old_stdout
 
 
-# def get_cltk_data_dir() -> str:
-#     """Defines where to look for the ``cltk_data`` dir.
-#     By default, this is located in a user's home directory
-#     and the directory is created there (``~/cltk_data``).
-#     However a user may customize where this goes with
-#     the OS environment variable ``$CLTK_DATA``. If the
-#     variable is found, then its value is used.
-#
-#     TODO: Run tests with a defined `$CLTK_DATA` environment variable)
-#
-#     >>> import os
-#     >>> del os.environ["CLTK_DATA"]
-#     >>> get_cltk_data_dir()
-#     abc
-#     >>> os.environ["CLTK_DATA"] = os.path.expanduser("~/cltk_data")
-#     abc
-#     >>>
-#     >>> os.environ["CLTK_DATA"] = os.path.expanduser("~/")
-#     """
-#     return __cltk_data_dir__
+def get_cltk_data_dir():
+    """Defines where to look for the ``cltk_data`` dir.
+    By default, this is located in a user's home directory
+    and the directory is created there (``~/cltk_data``).
+    However a user may customize where this goes with
+    the OS environment variable ``$CLTK_DATA``. If the
+    variable is found, then its value is used.
+
+    >>> from cltkv1.utils import get_cltk_data_dir
+    >>> import os
+    >>> os.environ["CLTK_DATA"] = os.path.expanduser("~/cltk_data")
+    >>> cltk_data_dir = get_cltk_data_dir()
+    >>> os.path.split(cltk_data_dir)[1]
+    'cltk_data'
+    >>> del os.environ["CLTK_DATA"]
+    >>> os.environ["CLTK_DATA"] = os.path.expanduser("~/custom_dir")
+    >>> cltk_data_dir = os.environ.get("CLTK_DATA")
+    >>> os.path.split(cltk_data_dir)[1]
+    'custom_dir'
+    """
+    import os  # pylint: disable=import-outside-toplevel
+
+    if "CLTK_DATA" in os.environ:
+        cltk_data_dir = os.path.expanduser(os.path.normpath(os.environ["CLTK_DATA"]))
+        if not os.path.isdir(cltk_data_dir):
+            raise FileNotFoundError(
+                "Custom data directory `%s` does not exist. "
+                "Update your OS environment variable `$CLTK_DATA` "
+                "or remove it." % cltk_data_dir
+            )
+        if not os.access(cltk_data_dir, os.W_OK):
+            raise PermissionError(
+                "Custom data directory `%s` must have "
+                "write permission." % cltk_data_dir
+            )
+    else:
+        cltk_data_dir = os.path.expanduser(os.path.normpath("~/cltk_data"))
+    return cltk_data_dir
+
+
+CLTK_DATA_DIR = get_cltk_data_dir()
